Move all the 'room' stuff into a package
